+++ CHANGELOG.md
@@ -1,17 +1,23 @@
 ## 0.4.0
 
 * 新增MyBottomMenu组件
+* 新增MyEndOfListWidget组件
+* 新增MyGroupBox组件
+* 新增MyList和MyCardList组件
+* 新增MyCard组件
 
 ## 0.3.0
 
 * 新增Splash页面
 * 新增MyDialog组件
 * 新增通过Key返回上一页或退出App功能
+* 新增MyApp.exit功能
 * 其他细节优化
+
 
 ## 0.2.0
 
-* 新增FocusableText组件
+* 新增Focus拓展
 * 新增MyButton组件
 * 新增MyMenuItem组件
 * 新增MyMenu组件
@@ -21,4 +27,4 @@
 
 ## 0.1.0
 
-* 新增MyApp.initialize组件
+* 新增MyApp.initialize功能